# perf: reduce the amount of URL parsing