**What's the problem this PR addresses?**

Parsing URLs is expensive

**How did you fix it?**

- Reuse URL instances by making some methods accept a `string | URL` and passing the parsed URL to them
- Avoid parsing known invalid URLs

**Result**

Tested on the Gatsby benchmark with a cached registry to avoid external network latency and a hot cache to only measure the time spent on resolution

```
YARN_IGNORE_PATH=1 hyperfine -w 1 --prepare="rm -rf yarn.lock .yarn" "node before.cjs --skip-builds" "node after.cjs --skip-builds"
Benchmark #1: node before.cjs --skip-builds
  Time (mean ± σ):     10.702 s ±  0.067 s    [User: 11.845 s, System: 1.473 s]
  Range (min … max):   10.631 s … 10.856 s    10 runs

Benchmark #2: node after.cjs --skip-builds
  Time (mean ± σ):     10.609 s ±  0.056 s    [User: 11.490 s, System: 1.525 s]
  Range (min … max):   10.521 s … 10.698 s    10 runs

Summary
  'node after.cjs --skip-builds' ran
    1.01 ± 0.01 times faster than 'node before.cjs --skip-builds'
```

**Checklist**
- [x] I have read the [Contributing Guide](https://yarnpkg.com/advanced/contributing).
- [x] I have set the packages that need to be released for my changes to be effective.
- [x] I will check that all automated PR checks pass before the PR gets reviewed.